% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm_build.R
\name{amm_set_W}
\alias{amm_set_W}
\title{Set the modulating basis on an AMM builder}
\usage{
amm_set_W(builder, W)
}
\arguments{
\item{builder}{An object of class \code{amm_builder}.}

\item{W}{A \code{W_basis} object, or \code{NULL} to disable the
modulating component.}
}
\value{
The modified \code{amm_builder}.
}
\description{
Store a \code{\link{W_basis}} object as the modulating component of
the specification under construction, or clear it by passing
\code{NULL}. The modulating component is global to all dimensions of
\eqn{\theta_i} and is therefore stored as a single top-level slot of
the builder.
}
\examples{
wb <- W_basis(type = "polynomial", degree = 2)
b <- amm_build(p = 1L) |>
  amm_set_a_uniform(~ x1) |>
  amm_set_W(wb)
print(b)

}
\seealso{
\code{\link{amm_build}}, \code{\link{W_basis}}
}
