% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpar_coef.R
\name{format.gdpar_coef}
\alias{format.gdpar_coef}
\title{One-line formatter for gdpar_coef objects}
\usage{
\method{format}{gdpar_coef}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gdpar_coef}.}

\item{...}{Unused.}
}
\value{
A length-1 character vector.
}
\description{
Useful for logs and condensed printouts.
}
