% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preflight_report.R
\name{format.gdpar_preflight_report}
\alias{format.gdpar_preflight_report}
\title{Format method for gdpar_preflight_report}
\usage{
\method{format}{gdpar_preflight_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_preflight_report}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Character scalar.
}
\description{
One-line representation of the report. Used by \code{format} in
contexts where a print is not desired.
}
