% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner_adapter.R
\name{is_gdpar_meta_learner_adapter}
\alias{is_gdpar_meta_learner_adapter}
\title{Test whether an object is a gdpar_meta_learner_adapter}
\usage{
is_gdpar_meta_learner_adapter(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} when \code{x} inherits from class
  \code{gdpar_meta_learner_adapter}, \code{FALSE} otherwise.
}
\description{
Test whether an object is a gdpar_meta_learner_adapter
}
\examples{
a <- gdpar_meta_learner_adapter(
  name = "dummy",
  fit_predict_fun = function(X, Y, T, X_newdata, level, seed_run) {
    list(cate_mean = rep(0, nrow(X_newdata)), cate_ci = NULL,
         state = NULL, notes = character(0))
  }
)
is_gdpar_meta_learner_adapter(a)
is_gdpar_meta_learner_adapter(list())
}
\seealso{
\code{\link{gdpar_meta_learner_adapter}}.
}
