% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_spec.R
\name{length.gdpar_formula_set}
\alias{length.gdpar_formula_set}
\title{Number of K-individual parameters in a gdpar_formula_set}
\usage{
\method{length}{gdpar_formula_set}(x)
}
\arguments{
\item{x}{An object of class \code{gdpar_formula_set}.}
}
\value{
Integer scalar equal to the number of slots.
}
\description{
Number of K-individual parameters in a gdpar_formula_set
}
