% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_spec.R
\name{names.gdpar_formula_set}
\alias{names.gdpar_formula_set}
\title{Slot names of a gdpar_formula_set}
\usage{
\method{names}{gdpar_formula_set}(x)
}
\arguments{
\item{x}{An object of class \code{gdpar_formula_set}.}
}
\value{
Character vector with the slot names in declaration order.
}
\description{
Slot names of a gdpar_formula_set
}
