% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm_spec.R
\name{print.amm_spec}
\alias{print.amm_spec}
\title{Print method for amm_spec objects}
\usage{
\method{print}{amm_spec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{amm_spec}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for amm_spec objects
}
