% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvm_check.R
\name{print.gdpar_bvm_report}
\alias{print.gdpar_bvm_report}
\title{Print method for gdpar_bvm_report objects}
\usage{
\method{print}{gdpar_bvm_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_bvm_report}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_bvm_report objects
}
