% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gdpar_diagnostics}
\alias{print.gdpar_diagnostics}
\title{Print method for gdpar_diagnostics objects}
\usage{
\method{print}{gdpar_diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_diagnostics}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_diagnostics objects
}
