% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.gdpar_family}
\alias{print.gdpar_family}
\title{Print method for gdpar_family objects}
\usage{
\method{print}{gdpar_family}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_family}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_family objects
}
