% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.gdpar_family_multi}
\alias{print.gdpar_family_multi}
\title{Print method for gdpar_family_multi objects}
\usage{
\method{print}{gdpar_family_multi}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_family_multi}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_family_multi objects
}
