% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_spec.R
\name{print.gdpar_formula_set}
\alias{print.gdpar_formula_set}
\title{Print method for gdpar_formula_set objects}
\usage{
\method{print}{gdpar_formula_set}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_formula_set}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_formula_set objects
}
