% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_bridge.R
\name{print.gdpar_geom_fit}
\alias{print.gdpar_geom_fit}
\title{Print method for gdpar_geom_fit objects}
\usage{
\method{print}{gdpar_geom_fit}(x, ...)
}
\arguments{
\item{x}{A \code{gdpar_geom_fit}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_geom_fit objects
}
