% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_laplace.R
\name{print.gdpar_geom_laplace}
\alias{print.gdpar_geom_laplace}
\title{Print method for gdpar_geom_laplace objects}
\usage{
\method{print}{gdpar_geom_laplace}(x, ...)
}
\arguments{
\item{x}{A \code{gdpar_geom_laplace}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_geom_laplace objects
}
