% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_orchestrator.R
\name{print.gdpar_geom_orchestration}
\alias{print.gdpar_geom_orchestration}
\title{Print method for gdpar_geom_orchestration objects}
\usage{
\method{print}{gdpar_geom_orchestration}(x, ...)
}
\arguments{
\item{x}{A \code{gdpar_geom_orchestration}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_geom_orchestration objects
}
