% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_engine.R
\name{print.gdpar_geom_target}
\alias{print.gdpar_geom_target}
\title{Print method for gdpar_geom_target objects}
\usage{
\method{print}{gdpar_geom_target}(x, ...)
}
\arguments{
\item{x}{A \code{gdpar_geom_target}.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_geom_target objects
}
