% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_suite.R
\name{print.gdpar_geometry_target}
\alias{print.gdpar_geometry_target}
\title{Print method for gdpar_geometry_target objects}
\usage{
\method{print}{gdpar_geometry_target}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_geometry_target}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_geometry_target objects
}
