% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_identifiability.R
\name{print.gdpar_identifiability_report}
\alias{print.gdpar_identifiability_report}
\title{Print method for gdpar_identifiability_report objects}
\usage{
\method{print}{gdpar_identifiability_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_identifiability_report}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_identifiability_report objects
}
