% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksd_joint.R
\name{print.gdpar_ksd_joint}
\alias{print.gdpar_ksd_joint}
\title{Print method for gdpar_ksd_joint}
\usage{
\method{print}{gdpar_ksd_joint}(x, digits = 4L, ...)
}
\arguments{
\item{x}{A \code{gdpar_ksd_joint} object.}

\item{digits}{Integer scalar; significant digits for the KSD value.
Defaults to 4.}

\item{...}{Reserved.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_ksd_joint
}
