% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner_adapter.R
\name{print.gdpar_meta_learner_adapter}
\alias{print.gdpar_meta_learner_adapter}
\title{Print method for gdpar_meta_learner_adapter}
\usage{
\method{print}{gdpar_meta_learner_adapter}(x, ...)
}
\arguments{
\item{x}{A \code{gdpar_meta_learner_adapter} object.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_meta_learner_adapter
}
