% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.gdpar_param_spec}
\alias{print.gdpar_param_spec}
\title{Print method for gdpar_param_spec objects}
\usage{
\method{print}{gdpar_param_spec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_param_spec}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_param_spec objects
}
\keyword{internal}
