% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{print.gdpar_prior}
\alias{print.gdpar_prior}
\title{Print method for gdpar_prior objects}
\usage{
\method{print}{gdpar_prior}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gdpar_prior}.}

\item{...}{Unused; present for S3 generic compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for gdpar_prior objects
}
