% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_eb_fb_methods.R
\name{print.summary.gdpar_eb_fb_comparison}
\alias{print.summary.gdpar_eb_fb_comparison}
\title{Print method for summary.gdpar_eb_fb_comparison}
\usage{
\method{print}{summary.gdpar_eb_fb_comparison}(x, digits = 3L, ...)
}
\arguments{
\item{x}{A \code{summary.gdpar_eb_fb_comparison} object.}

\item{digits}{Integer scalar passed to \code{format()}; defaults to
3.}

\item{...}{Unused.}
}
\value{
The object \code{x} invisibly.
}
\description{
Print method for summary.gdpar_eb_fb_comparison
}
