% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_spec.R
\name{[[.gdpar_formula_set}
\alias{[[.gdpar_formula_set}
\title{Extract a single formula from a gdpar_formula_set}
\usage{
\method{[[}{gdpar_formula_set}(x, i)
}
\arguments{
\item{x}{An object of class \code{gdpar_formula_set}.}

\item{i}{A character scalar with the slot name or an integer index.}
}
\value{
The formula stored at slot \code{i}.
}
\description{
Extract a single formula from a gdpar_formula_set
}
